storage: Make initial "at_boot" values consistent
The default is now determined by at_boot_input. Previously, btrfs
subvolumes would be created with "local" instead of the desired
"nofail" default.

Also, factor out the updating mechanism, and make sure the subvolume
creation dialog uses it as well.
